pyproject.toml
[project]
name = "iaEditais"
version = "1.0.0"
description = ""
authors = [
    {name = "gmcosta",email = "geu_costa@outlook.com"}
]
license = {text = "GNU AFFERO GENERAL PUBLIC LICENSE"}
readme = "README.md"
requires-python = ">=3.12,<4.0" 
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "testcontainers[postgres] (>=4.10.0,<5.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "langchain (==0.3.23)",
    "langchain-community (==0.3.20)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "langchain-postgres (>=0.0.13,<0.0.14)",
    "langchain-ollama (>=0.3.0,<0.4.0)",
    "langchain-google (>=0.1.1,<0.2.0)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "langchain-deepseek (>=0.1.3,<0.2.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-mermaid2-plugin (>=1.2.1,<2.0.0)",
    "mkdocs-material (>=9.6.14,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
]

[tool.poetry]
packages = [{ include = "iaEditais" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.0"
taskipy = "^1.14.1"
ruff = "^0.11.4"
fpdf = ">=1.7.2,<2.0.0"

[tool.ruff]
line-length = 81
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check --fix --unsafe-fixes; ruff format'
run = 'fastapi dev iaEditais/app.py'
test = 'task lint; pytest -s -x --lf --cov=iaEditais -vv; coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
